perf: avoid unnecessary Android viewport probes
