+++ src/platforms/android/touch-executor.ts
@@ -1,9 +1,6 @@
 import type { DeviceInfo } from '../../kernel/device.ts';
 import type { Rect } from '../../kernel/snapshot.ts';
-import {
-  resolveAndroidGestureViewportProvider,
-  resolveAndroidTouchInjector,
-} from './adb-executor.ts';
+import { resolveAndroidTouchProvider } from './adb-executor.ts';
 import {
   executeAndroidMultiTouchHelperPlan,
   readAndroidMultiTouchHelperViewport,
@@ -15,16 +12,23 @@ export async function executeAndroidTouchPlan(
   device: DeviceInfo,
   plan: AndroidTouchPlan,
 ): Promise<Record<string, unknown>> {
-  const providerTouch = resolveAndroidTouchInjector(device);
-  if (providerTouch) {
-    const result = (await providerTouch(plan)) ?? {};
+  const provider = resolveAndroidTouchProvider(device);
+  if (provider) {
+    const providerPlan =
+      plan.intent === 'longPress'
+        ? {
+            ...plan,
+            viewport: validateAndroidGestureViewport(await provider.gestureViewport()),
+          }
+        : plan;
+    const result = (await provider.touch(providerPlan)) ?? {};
     return { backend: 'provider-native-touch', ...result };
   }
   return await executeAndroidMultiTouchHelperPlan(device, plan);
 }
 
 export async function readAndroidGestureViewport(device: DeviceInfo): Promise<Rect> {
-  const providerViewport = resolveAndroidGestureViewportProvider(device);
-  if (providerViewport) return validateAndroidGestureViewport(await providerViewport());
+  const provider = resolveAndroidTouchProvider(device);
+  if (provider) return validateAndroidGestureViewport(await provider.gestureViewport());
   return await readAndroidMultiTouchHelperViewport(device);
 }
